backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
astroid==3.3.9
attrs==25.3.0
black==25.1.0
boto3==1.37.28
botocore==1.37.28
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
dill==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-phone-verify==3.1.0
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
drf-spectacular==0.28.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==6.0.1
itsdangerous==2.2.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
multidict==6.3.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
phonenumbers==9.0.2
phonenumberslite==9.0.2
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
s3transfer==0.11.4
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomlkit==0.13.2
twilio==9.5.1
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
